+++ request.go
@@ -7,6 +7,8 @@ import (
 	"net/url"
 	"time"
 
+	"github.com/System-Glitch/goyave/v2/cors"
+
 	"github.com/System-Glitch/goyave/v2/helper/filesystem"
 	"github.com/System-Glitch/goyave/v2/validation"
 	"github.com/google/uuid"
@@ -16,6 +18,7 @@ import (
 // Contains the validated body in the Data attribute if the route was defined with a request generator function
 type Request struct {
 	httpRequest *http.Request
+	corsOptions *cors.Options
 	cookies     []*http.Cookie
 	Rules       validation.RuleSet
 	Data        map[string]interface{}
@@ -91,6 +94,18 @@ func (r *Request) Referrer() string {
 // UserAgent returns the client's User-Agent, if sent in the request.
 func (r *Request) UserAgent() string {
 	return r.httpRequest.UserAgent()
+}
+
+// CORSOptions returns the CORS options applied to this request, or nil.
+// The returned object is a copy of the options applied to the router.
+// Therefore, altering the returned object will not alter the router's options.
+func (r *Request) CORSOptions() *cors.Options {
+	if r.corsOptions == nil {
+		return nil
+	}
+
+	cpy := *r.corsOptions
+	return &cpy
 }
 
 // Has check if the given field exists in the request data.